SSCS-2648 Refactor CaseLoaderService
Autowire services in constructor to follow approach with rest classes.
Sketch the service calls needed to make the App work.
Move CaseLoaderService to services root directory.
Move JschChannel bean configuration to the config as for other classes.
